LPS-63470 move modules
