+++ modules/apps/business-productivity/portal-workflow/portal-workflow-kaleo-definition-api/src/main/java/com/liferay/portal/workflow/kaleo/definition/export/NodeExporter.java
@@ -12,18 +12,16 @@
  * details.
  */
 
-package com.liferay.portal.workflow.kaleo.export;
+package com.liferay.portal.workflow.kaleo.definition.export;
 
-import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.xml.Element;
+import com.liferay.portal.workflow.kaleo.definition.Node;
 
 /**
  * @author Michael C. Han
  */
-public interface DefinitionExporter {
+public interface NodeExporter {
 
-	public String export(long kaleoDefinitionId) throws PortalException;
-
-	public String export(long companyId, String name, int version)
-		throws PortalException;
+	public void exportNode(Node node, Element element, String namespace);
 
 }
